Misc Refactors - Fix several issues relating to Vulkan implementation - Disabled warnings from external headers and enabled -wpedantic - Set up Conan script for finding the Vulkan SDK - Set up remote for hosting aforementioned Conan script - Work on getting the engine to build with MSVC. Currently awaiting   compiler bugfix: https://developercommunity.visualstudio.com/t/Keyword-requires-within-if-statement-d/1287202 - Minor refactors within game code - Implement basic testing framework - Implement tests for event system - Implement spatial event dispatcher

+++ VoxelEngine/core/typedefs/constants.hpp
@@ -9,8 +9,8 @@ namespace ve {
     constexpr inline double sqrt_2 = std::numbers::sqrt2;
     constexpr inline double sqrt_3 = std::numbers::sqrt3;
     
-    constexpr inline float pi_f     = std::numbers::pi;
-    constexpr inline float tau_f    = 2 * std::numbers::pi;
-    constexpr inline float sqrt_2_f = std::numbers::sqrt2;
-    constexpr inline float sqrt_3_f = std::numbers::sqrt3;
+    constexpr inline float pi_f     = float(std::numbers::pi);
+    constexpr inline float tau_f    = float(2 * std::numbers::pi);
+    constexpr inline float sqrt_2_f = float(std::numbers::sqrt2);
+    constexpr inline float sqrt_3_f = float(std::numbers::sqrt3);
 }
